N-API mapnik.VectorTile (WIP)
